changed mediumint back to int for database creation

+++ sql/init.sql
@@ -1,5 +1,5 @@
 CREATE TABLE Users (
-    userid mediumint NOT NULL AUTO_INCREMENT,
+    userid INT NOT NULL AUTO_INCREMENT,
     username varchar(50) NOT NULL,
     email varchar(50) NOT NULL,
     hash varchar(100) NOT NULL,
@@ -15,7 +15,7 @@ CREATE TABLE Chats (
 );
 
 CREATE TABLE ChatMembers (
-    userid mediumint NOT NULL,
+    userid INT NOT NULL,
     chatid INT NOT NULL,
     FOREIGN KEY(userid) REFERENCES Users(userid),
     FOREIGN KEY(chatid) REFERENCES Chats(chatid),
@@ -25,7 +25,7 @@ CREATE TABLE ChatMembers (
 -- Shouldn't be set to GroupMembers to avoid foregin key join duplication
 CREATE TABLE ChatMessages(
     messageid INT NOT NULL AUTO_INCREMENT,
-    userid mediumint NOT NULL,
+    userid INT NOT NULL,
     chatid INT NOT NULL,
     message VARCHAR(2000) NOT NULL,
     timestamp DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -36,8 +36,8 @@ CREATE TABLE ChatMessages(
 
 CREATE TABLE UserFriends(
     connectionid INT NOT NULL AUTO_INCREMENT,
-    userid mediumint NOT NULL,
-    friendid mediumint NOT NULL,
+    userid INT NOT NULL,
+    friendid INT NOT NULL,
     accepted BIT NOT NULL DEFAULT 0,
     PRIMARY KEY (connectionid),
     FOREIGN KEY(userid) REFERENCES Users(userid),
@@ -45,8 +45,8 @@ CREATE TABLE UserFriends(
 );
 
 CREATE TABLE UserBlocks(
-    userid mediumint NOT NULL,
-    blockid mediumint NOT NULL,
+    userid INT NOT NULL,
+    blockid INT NOT NULL,
     FOREIGN KEY(userid) REFERENCES Users(userid),
     FOREIGN KEY(blockid) REFERENCES Users(userid),
     PRIMARY KEY(userid, blockid)
